+++ db/services.php
@@ -15,7 +15,7 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Plugin version details.
+ * External service definitions for local_smartnoticespro.
  *
  * @package   local_smartnoticespro
  * @copyright 2026 Jesus Antonio Jimenez Aviña <antoniomexdf@gmail.com> <antoniojamx@gmail.com>
@@ -24,9 +24,14 @@
 
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->component = 'local_smartnoticespro';
-$plugin->version = 2026032300;
-$plugin->requires = 2024100700; // Moodle 4.5.
-$plugin->supported = [405, 405];
-$plugin->maturity = MATURITY_STABLE;
-$plugin->release = '0.2.34-pro';
+$functions = [
+    'local_smartnoticespro_track_notice_event' => [
+        'classname' => 'local_smartnoticespro\\external\\track_notice_event',
+        'methodname' => 'execute',
+        'classpath' => '',
+        'description' => 'Tracks Smart Notices Pro modal interactions using Moodle external services.',
+        'type' => 'write',
+        'ajax' => true,
+        'capabilities' => 'local/smartnoticespro:viewnotices',
+    ],
+];